refactor(ai): topic-split S7D findings and orchestrator test support
Collapse the 1667-line findings monolith and 794-line orchestrator
support library into topical modules + thin barrels for Phase-5 AC3/AC4
(Refs #3972), without changing AI semantics.
